+++ packages/xy/tests/scatter/ScatterLabel.test.tsx
@@ -2,13 +2,17 @@ import { render, screen } from '@testing-library/react'
 import { Chart } from '@chsk/core'
 import { Scatter, ScatterLabel } from '../../src/scatter'
 import { scatterProps } from './scatter.props'
+import { getTransform } from '../../../core/tests/utils'
+
+// tests use a dataset with two series: 'linear' (y=x) and 'quadratic' (y=x**2)
+// the x domain for both series is [1, 8]
 
 describe('ScatterLabel', () => {
     it('creates a label', () => {
         render(
             <Chart>
                 <Scatter {...scatterProps}>
-                    <ScatterLabel id={'linear'} x={4}>
+                    <ScatterLabel id={'linear'} position={4}>
                         Label
                     </ScatterLabel>
                 </Scatter>
@@ -25,22 +29,43 @@ describe('ScatterLabel', () => {
         render(
             <Chart size={[400, 300]} padding={[0, 0, 0, 0]}>
                 <Scatter {...scatterProps}>
-                    <ScatterLabel id={'linear'} x={1} units={'relative'}>
+                    <ScatterLabel id={'quadratic'} position={1} positionUnits={'relative'}>
+                        Label
+                    </ScatterLabel>
+                </Scatter>
+            </Chart>
+        )
+        // x=1 in view-relative units means at x=400 in svg units,
+        // the closest data point will be [8, 64], which is near top-right corner
+        const text = screen.getByRole('scatter-label').querySelector('text')
+        const g = text?.closest('g')
+        expect(getTransform(g ?? text, 'X')).toEqual(400)
+        expect(getTransform(g ?? text, 'Y')).toBeLessThan(50)
+    })
+
+    it('creates a label using xy position', () => {
+        render(
+            <Chart size={[400, 300]} padding={[0, 0, 0, 0]}>
+                <Scatter {...scatterProps}>
+                    <ScatterLabel id={'quadratic'} position={[8, 50]} positionUnits={'view'}>
                         Label
                     </ScatterLabel>
                 </Scatter>
             </Chart>
         )
         const text = screen.getByRole('scatter-label').querySelector('text')
-        // x=1 in view-relative units means at x=400 in svg units
-        expect(text?.closest('g')?.getAttribute('style')).toContain('translateX(400')
+        // data has x=1..8 and y=1..64
+        // the nearest data point to [8, 50] should be [7, 49] rather than [8, 64]
+        const g = text?.closest('g')
+        expect(getTransform(g ?? text, 'X')).toBeLessThan(400)
+        expect(getTransform(g ?? text, 'Y')).toBeGreaterThan(50)
     })
 
     it('skips work for non-existent series', () => {
         render(
             <Chart>
                 <Scatter {...scatterProps}>
-                    <ScatterLabel id={'non-existent'} x={4}>
+                    <ScatterLabel id={'non-existent'} position={4}>
                         Label
                     </ScatterLabel>
                 </Scatter>
@@ -53,7 +78,7 @@ describe('ScatterLabel', () => {
         render(
             <Chart>
                 <Scatter {...scatterProps}>
-                    <ScatterLabel id={'linear'} x={4.2} autoRotate={true}>
+                    <ScatterLabel id={'linear'} position={4.2} autoRotate={true}>
                         Label
                     </ScatterLabel>
                 </Scatter>
@@ -69,7 +94,7 @@ describe('ScatterLabel', () => {
         render(
             <Chart>
                 <Scatter {...scatterProps}>
-                    <ScatterLabel id={'linear'} x={4.8} autoRotate={true}>
+                    <ScatterLabel id={'linear'} position={4.8} autoRotate={true}>
                         Label
                     </ScatterLabel>
                 </Scatter>
@@ -85,7 +110,7 @@ describe('ScatterLabel', () => {
         render(
             <Chart data={{ disabledKeys: new Set<string>(['linear']) }}>
                 <Scatter {...scatterProps}>
-                    <ScatterLabel id={'linear'} x={4}>
+                    <ScatterLabel id={'linear'} position={4}>
                         Label
                     </ScatterLabel>
                 </Scatter>
@@ -104,7 +129,7 @@ describe('ScatterLabel', () => {
         render(
             <Chart>
                 <Scatter {...scatterProps} data={emptyData}>
-                    <ScatterLabel id={'empty'} x={4}>
+                    <ScatterLabel id={'empty'} position={4}>
                         Label
                     </ScatterLabel>
                 </Scatter>
@@ -127,7 +152,7 @@ describe('ScatterLabel', () => {
         render(
             <Chart>
                 <Scatter {...scatterProps} data={verticalData}>
-                    <ScatterLabel id={'vertical'} x={0.5} autoRotate={true}>
+                    <ScatterLabel id={'vertical'} position={0.5} autoRotate={true}>
                         Label
                     </ScatterLabel>
                 </Scatter>
@@ -146,17 +171,24 @@ describe('ScatterLabel', () => {
                 data: [{ x: 1, y: 2 }],
             },
         ]
+        // match single-value and xy-value positions to data points
         render(
             <Chart>
                 <Scatter {...scatterProps} data={singleData}>
-                    <ScatterLabel id={'single'} x={0.5} autoRotate={true}>
+                    <ScatterLabel id={'single'} position={0.5} autoRotate={true}>
+                        Label
+                    </ScatterLabel>
+                    <ScatterLabel id={'single'} position={[3, 4]} autoRotate={true}>
                         Label
                     </ScatterLabel>
                 </Scatter>
             </Chart>
         )
-        const text = screen.queryByRole('scatter-label')?.querySelector('text')
-        expect(text?.textContent).toBe('Label')
-        expect(text?.closest('g')?.getAttribute('style')).toContain('rotate(0')
+        const result = screen.queryAllByRole('scatter-label')
+        result.map(label => {
+            const text = label.querySelector('text')
+            expect(text?.textContent).toBe('Label')
+            expect(text?.closest('g')?.getAttribute('style')).toContain('rotate(0')
+        })
     })
 })